Added the "Edit" button the ScenarioListActivity, as well as 2 classes (ScenarioDetailEditFragment.java and ScenarioDetailEditActivity.java).
The 2 new classes are to be used for making a list view once the user
presses the "Edit" button. This list view will contain the list of
questions from the database, as well as a list of text fields containing
the corresponding answers to each question. The answers will be
editable, and once an update button is pushed, they will update the
answer values of the database.